fixing healthbar

+++ Assets/HealthBar.cs
@@ -18,7 +18,7 @@ public class HealthBar : MonoBehaviour
 		fill.color = gradient.Evaluate(1f);
 	}
 
-    public void SetHealth(float health)
+    public void SetHealth(int health)
 	{
 		slider.value = health;
